refactor(services): split orchestrator.rs into submodules
Extract PostgreSQL auto-start helpers and shard lookup into private
sibling modules; orchestrator.rs now focuses on service lifecycle.

Public API unchanged. No behavior changes.
