+++ crates/services/src/lib.rs
@@ -15,3 +15,5 @@ pub mod database;
 pub mod mercury;
 pub mod minigame;
 pub mod orchestrator;
+mod orchestrator_postgres;
+mod orchestrator_shards;
